# apps/api/openapi.gen.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Char AI API",
    "description": "AI services API for speech-to-text transcription, LLM chat completions, and subscription management",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/subscription/delete-account": {
      "delete": {
        "tags": [
          "subscription"
        ],
        "operationId": "delete_account",
        "responses": {
          "202": {
            "description": "Account deletion durably accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Durable account cleanup unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteAccountResponse": {
        "type": "object",
        "required": [
          "deleted",
          "pending"
        ],
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "pending": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token"
      }
    }
  },
...
}